docmatcher/backend/pyproject.toml
[tool.poetry]
name = "docmatcher"
version = "0.1.0"
description = ""
authors = ["nownabe"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
fastapi = "^0.90.1"
protobuf = "^4.21.12"
google-cloud-spanner = "^3.27.1"
google-cloud-pubsub = "^2.14.1"
mecab-python3 = "^1.0.6"
gensim = "^4.3.0"
google-cloud-storage = "^2.7.0"
grpcio = "^1.51.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
